Rename WalletListModal2 to WalletListModal
- Rename CryptoExchangeCreateWalletRow to WalletListModalCreateRow,
added types on it and optimize some code
